folder and file for repeat function
