+++ pages/visualizations/overview/total_contributor_growth.py
@@ -8,6 +8,7 @@ import pandas as pd
 import datetime as dt
 import logging
 import plotly.express as px
+from utils.graph_utils import get_graph_time_values
 
 gc_total_contributor_growth = dbc.Card(
     [
@@ -169,7 +170,7 @@ def contributor_growth_bar_graph(df_contrib, bin_size):
         group = df_contrib.groupby(pd.Grouper(key="created_at", axis=0, freq="1Y")).size()
 
     # time values for graph
-    x_r, x_name, hover = get_graph_time_values(bin_size)
+    x_r, x_name, hover, period = get_graph_time_values(bin_size)
 
     # reset index from group-by aggregation step
     group = group.reset_index()
@@ -267,30 +268,3 @@ def contributor_growth_line_bar(df_contrib):
         margin_r=20,
     )
     return fig
-
-
-def get_graph_time_values(interval):
-    # helper values for building graph
-    today = dt.date.today()
-    x_r = None
-    x_name = "Year"
-    hover = "Year: %{x|%Y}"
-
-    # graph input values based on date interval selection
-    if interval == 86400000:  # if statement for days
-        x_r = [str(today - dt.timedelta(weeks=4)), str(today)]
-        x_name = "Day"
-        hover = "Day: %{x|%b %d, %Y}"
-    elif interval == "D1":
-        x_r = [str(today - dt.timedelta(weeks=4)), str(today)]
-        x_name = "Day"
-        hover = "Day: %{x|%b %d, %Y}"
-    elif interval == 604800000:  # if statmement for weeks
-        x_r = [str(today - dt.timedelta(weeks=30)), str(today)]
-        x_name = "Week"
-        hover = "Week: %{x|%b %d, %Y}"
-    elif interval == "M1":  # if statement for months
-        x_r = [str(today - dt.timedelta(weeks=104)), str(today)]
-        x_name = "Month"
-        hover = "Month: %{x|%b %Y}"
-    return x_r, x_name, hover
