+++ utils/graph_utils.py
@@ -1,0 +1,38 @@
+from dash import html, dcc
+import dash
+import dash_bootstrap_components as dbc
+from dash import callback
+from dash.dependencies import Input, Output, State
+import plotly.graph_objects as go
+import pandas as pd
+import datetime as dt
+import logging
+import numpy as np
+import plotly.express as px
+
+
+def get_graph_time_values(interval):
+    # helper values for building graph
+    today = dt.date.today()
+    x_r = None
+    x_name = "Year"
+    hover = "Year: %{x|%Y}"
+    period = "Y"
+
+    # graph input values based on date interval selection
+    if interval == 86400000:  # if statement for days
+        x_r = [str(today - dt.timedelta(weeks=4)), str(today)]
+        x_name = "Day"
+        hover = "Day: %{x|%b %d, %Y}"
+        period = "D"
+    elif interval == 604800000:  # if statmement for weeks
+        x_r = [str(today - dt.timedelta(weeks=30)), str(today)]
+        x_name = "Week"
+        hover = "Week: %{x|%b %d, %Y}"
+        period = "W"
+    elif interval == "M1":  # if statement for months
+        x_r = [str(today - dt.timedelta(weeks=104)), str(today)]
+        x_name = "Month"
+        hover = "Month: %{x|%b %Y}"
+        period = "M"
+    return x_r, x_name, hover, period
